% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_numeric_factor}
\alias{as_numeric_factor}
\title{Converts factor to number}
\usage{
as_numeric_factor(x)
}
\arguments{
\item{x}{A factor}
}
\value{
A numeric with factor levels??
}
\description{
Converts factor to number
}
\examples{
str(ToothGrowth)
x <- ToothGrowth$supp
}
