% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{avg_dataframe}
\alias{avg_dataframe}
\title{Functions defined in this file:
avg_dataframe
avg_dataset
Average a dataframe with continuous and discrete variables columnwise.}
\usage{
avg_dataframe(
  df,
  group_by = 1,
  rows_to_keep = NULL,
  cols_to_keep = NULL,
  sort_cols = FALSE
)
}
\arguments{
\item{df}{A data frame to average across. Note: Only numeric columns will be averaged, while discrete columns will retain the first instance for each group.}

\item{group_by}{Name or column of index to average valuesby.}

\item{rows_to_keep}{Numeric, logical or character vectors of rows to keep in subset (length must be number of rows of df). Note: character for named rows only}

\item{cols_to_keep}{Numeric, logical or character vectors of columns to keep in subset (length must be number of columns of df). Note: character for named columns only}

\item{sort_cols}{logical, should we sort columns alphabetically?}
}
\value{
A data frame averaged column by column.
}
\description{
Average a dataframe with continuous and discrete variables columnwise.
}
