% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{avg_dataset}
\alias{avg_dataset}
\title{Averages a dataset object.}
\usage{
avg_dataset(
  ds,
  group_by = 2,
  new_name = "Averaged",
  rows_to_keep = NULL,
  cols_to_keep = NULL
)
}
\arguments{
\item{ds}{A dataset object (a list with any of vals, rowAnn, colAnn).}

\item{group_by}{Name or index of column in ds$rowAnn to average values by.}

\item{rows_to_keep}{Numeric, logical or character vectors of rows to keep in ds$vals (length must be number of rows of ds$vals). Note: character if rows are named.}

\item{cols_to_keep}{Numeric, logical or character vectors of columns to keep in ds$vals(length must be number of columns of ds$vals). Note: character if columns are named.}

\item{sort_cols}{logical, should we sort columns alphabetically?}
}
\value{
A dataset subsetted specified rows/columns and averaged.
}
\description{
Averages a dataset object.
}
