% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bin_vars}
\alias{bin_vars}
\title{Replaces values in a vector that equal to certain values to a new value}
\usage{
bin_vars(v, orig_vals, new_val)
}
\description{
@param v A vector
@param orig_vals A vector of values in v to replace
@param new_val The value to replace original values with
@return A vector like v, except the orig_vals have been replaced with new_val
@export
}
