% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_corrmat}
\alias{cluster_corrmat}
\title{Re-orders a correlation matrix by hierarchical clustering}
\usage{
cluster_corrmat(mat, dist_method = "euclidean", link_method = "complete")
}
\arguments{
\item{mat}{Correlation matrix or data frame}

\item{dist_method}{The distance measure to be used. Accepts the same values as method parameter in ?dist.}

\item{link_method}{The agglomeration/clustering method to be used. Accepts the same values as method parameter in ?hclust.}
}
\value{
An ordered matrix based on hierarchical clustering
}
\description{
Re-orders a correlation matrix by hierarchical clustering
}
