% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_df}
\alias{cluster_df}
\title{Functions defined in this file:
cluster_df
cluster_corrmat
Re-orders a matrix by hierarchical clustering}
\usage{
cluster_df(
  df3,
  dist_method = "euclidean",
  link_method = "ward.D",
  clust.by.row = T
)
}
\arguments{
\item{df3}{Numeric matrix or data frame}

\item{dist_method}{The distance measure to be used. Accepts the same values as method parameter in ?dist.}

\item{link_method}{The agglomeration/clustering method to be used. Accepts the same values as method parameter in ?hclust.}

\item{clust.by.row}{Logical indicating whether to cluster by row, else columnwise.}
}
\value{
An ordered matrix based on hierarchical clustering
}
\description{
Functions defined in this file:
cluster_df
cluster_corrmat
Re-orders a matrix by hierarchical clustering
}
