% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap_plots.R
\name{create_heatmap_plots}
\alias{create_heatmap_plots}
\title{Create heatmap plots helper in a folder in streamlined approach}
\usage{
create_heatmap_plots(
  ds,
  rowAnns = c(1, NA),
  colAnns = c(1, NA),
  out_dir = ".",
  labels = "Sub1"
)
}
\arguments{
\item{out_dir}{The output directory where the plots will be saved, default is current working directory.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}
}
\description{
Run heatmap analysis helper
}
