% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap_plots.R
\name{create_heatmap_plots_helper}
\alias{create_heatmap_plots_helper}
\title{Create heatmap plots helper}
\usage{
create_heatmap_plots_helper(
  ds,
  rowAnns = 1,
  colAnns = NA,
  out_dir = ".",
  labels = "",
  pal = NA,
  clust_row = F,
  clust_col = F,
  make.heatmap = F,
  make.corrplot = F,
  make.overview.corrscatt = F,
  make.indiv.corrscatt = F,
  make.barplot = F,
  make.indiv.boxplot = F,
  make.overview.boxplot = F,
  make.FC.pval.plot = T
)
}
\arguments{
\item{pal}{A named character vector of colors (R or hex), where the names are the groups in row annotations and column annotations}

\item{clust_row, clust_col}{Logicals indicating whether to cluster rows and columns in heatmap.}

\item{make.heatmap, make.corrplot, make.overview.corrscatt, make.indiv.corrscatt, make.barplot, make.indiv.boxplot, make.overview.boxplot, make.FC.pval.plot}{Logicals indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}
}
\description{
Create heatmap plots helper
}
