% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plots.R
\name{create_plots}
\alias{create_plots}
\title{Create plots in a folder in streamlined approach}
\usage{
create_plots(
  ds,
  rowAnns = 1,
  colAnns = NA,
  out_dir = ".",
  labels = "Sub1",
  var_colors = NULL,
  gradient_palette = NULL,
  corr_method = "pearson",
  pval.test = "wilcox.test",
  pval.label = "p.signif",
  make.indiv.boxplot = F,
  make.overview.boxplot = F,
  make.heatmap = F,
  make.corrplot = F,
  make.overview.corrscatt = F,
  make.indiv.corrscatt = F,
  make.barplot = F,
  make.FC.pval.plot = F,
  also.complete = F
)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn. c(MainComparison, Subgroup)}

\item{colAnns}{A character vector of 1-2 column names in ds$colAnn. c(Parameter, Feature/Stain/Gene)}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{var_colors}{A named vector with colors as values and annotations/groups as names.}

\item{gradient_palette}{RColorBrewer palette name for gradients (e.g. heatmap, correlation plots). See RColorBrewer::display.brewer.all() for all options.}

\item{corr_method}{Method for correlation (one of "pearson","spearman","kendall").}

\item{pval.test}{Which two-samples testing should be used? String corresponding to "method" parameter in \code{\link[ggpubr]{stat_compare_means}}. Allowed values are "t.test" and "wilcox.test".}

\item{pval.label}{How to display p-values? String corresponding to "label" parameter in \code{\link[ggpubr]{stat_compare_means}}. Allowed values are "p.signif" (stars) and "p.format" (number).}

\item{make.indiv.boxplot}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{make.overview.boxplot}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{make.heatmap}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{make.corrplot}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{make.overview.corrscatt}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{make.indiv.corrscatt}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{make.barplot}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{make.FC.pval.plot}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{also.complete}{Logical indicating whether to also make "complete cores" plots as a seperate folder, default FALSE.}
}
\description{
Create plots in a folder in streamlined approach
}
