% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_numeric}
\alias{df_to_numeric}
\title{Coerce a data.frame to a numeric.data.frame}
\usage{
df_to_numeric(df, row.or.col = 2)
}
\arguments{
\item{df}{Numeric data frame or matrix}

\item{row.or.col}{A number indicating direction to apply function: either 1 for row-wise, 2 (default) for column-wise}
}
\value{
A data frame in which all columns/rows are numeric
}
\description{
Coerce a data.frame to a numeric.data.frame
}
