% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{evaluate_filter}
\alias{evaluate_filter}
\title{Subsets a dataframe based on filters}
\usage{
evaluate_filter(df, filt)
}
\arguments{
\item{df}{A dataframe}

\item{filters}{A string with 3 parts: 1) column name in df, 2) operator either != or ==, 3) value in column to exclude/include}
}
\description{
Subsets a dataframe based on filters
}
\details{
evaluate_filter(df, "Smoker==Yes") # positively select for smokers
}
