% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects_run_hourglass.R
\name{get_colors}
\alias{get_colors}
\title{Get color palette from input Excel file.}
\usage{
get_colors(file_xl, sheet = "Colors")
}
\arguments{
\item{file_xl}{Excel file (.xlsx) containing user options for hourglass run.}

\item{sheet}{Name of worksheet in file_xl containing table where column 1 is variables and column 2 is hex code}
}
\value{
List of colors, where elements are hex codes and element names are variables. As a result of "cleaning", element names don't have 2 parts, are duplicated, and is white e.g. list("Tumour"="#2f4f4Fff", "Stroma"="#d2691eff")
}
\description{
Get color palette for all rowAnn variables for use by run_hourglass function from input Excel file.
}
