% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{get_comparison_name}
\alias{get_comparison_name}
\title{Customize name of comparison / output folder depending on inclusion/exclusion criteria and presence of row annotation 2}
\usage{
get_comparison_name(
  current,
  filters,
  delim = ";",
  all_out_dirs = NULL,
  rowAnn2 = NA
)
}
\arguments{
\item{filters}{A string in the form of filters delimited by default ";". Each filter has 3 parts: 1) column name in df, 2) operator either != or ==, 3) value in column to exclude/include}

\item{delim}{A string/character to seperate individuals filter by, default is ";"}

\item{all_out_dirs}{A character vector of directory names, if the current_dir name exists, the algorithm will append "1", "2", and so on until the dir name is unique.}

\item{rowAnn2}{Name of row annotation 2 if applicable, otherwise NA.}

\item{current_dir}{The name of the output directory}
}
\value{
New name of output folder
}
\description{
Customize name of comparison / output folder depending on inclusion/exclusion criteria and presence of row annotation 2
}
