% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects_run_Hourglass.R
\name{get_comparisons}
\alias{get_comparisons}
\title{Get comparisons to run from input Excel file.}
\usage{
get_comparisons(file_xl, sheet = "Comparisons")
}
\arguments{
\item{file_xl}{Excel file (.xlsx) containing user options for hourglass run.}

\item{sheet}{Name of worksheet in file_xl that contains comparisons to run.}
}
\value{
A data frame containing comparisons to run compatible with run_hourglass function (i.e. column names are options, rows are each comparison to run)
}
\description{
Get comparisons to run data frame for run_hourglass function from input Excel file.
}
