% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customAn_ds.R
\name{get_customAn_ds}
\alias{get_customAn_ds}
\title{Subset dataset for custom analysis}
\usage{
get_customAn_ds(ds, customAn, i, new_colAnns = NA, customAn_order = T)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{customAn}{A list of 2 data frames for plotting specific rows and columns. @seealso customAn read in using \code{\link{import_custom_analysis_file}}, see description of file.}

\item{i}{Row number in customAn$keys.}

\item{new_colAnns}{Name of new column annotations in custom annotations file}

\item{customAn_order}{Logical, should we preserve customAn order?}
}
\value{
A list object specifying: ds, colAnns, customAn_name (name of custom analysis)
}
\description{
Get a subset of dataset list object from custom analysis table
}
\examples{
results <- get_customAn_ds(ds, customAn, i = 1)
}
