% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects_run_hourglass.R
\name{get_datasets}
\alias{get_datasets}
\title{Prepare dataset objects in run_hourglass function.}
\usage{
get_datasets(comparisons, datasets = NULL)
}
\arguments{
\item{comparisons}{A data frame containing comparisons to run compatible with run_hourglass function (i.e. column names are options, rows are each comparison to run). See ?get_comparisons for more info.}

\item{datasets}{Optional. List of 2 elements 1) samples, 2) patients, which are the dataset objects for BySample and ByPatient analysis respectively.}
}
\value{
Returns a list of 4 dataset objects with names: "samples", "samples_imp", "patients", "patients_imp"
}
\description{
Prepare dataset objects: samples, patients (and imputed datasets) based on user options.#'
}
