% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_duplicated_cases}
\alias{get_duplicated_cases}
\title{Subset data frame to rows that are duplicated at least once}
\usage{
get_duplicated_cases(d, col, rm.NA = NA)
}
\arguments{
\item{d}{A data frame}

\item{col}{Column in d that you want duplicates from}

\item{rm.NA}{Column you want to remove duplicates or zeros from prior to finding duplication (optional)}
}
\value{
A data frame in which only duplicated rows of col exist
}
\description{
Subset data frame to rows that are duplicated at least once
}
