% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_sets.R
\name{get_feat_sets_ds}
\alias{get_feat_sets_ds}
\title{Subset dataset for custom analysis}
\usage{
get_feat_sets_ds(ds, feat_sets, i, new_colAnns = NA, feat_sets_order = T)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{feat_sets}{A list of 2 data frames for plotting specific rows and columns.}

\item{i}{Row number in feat_sets$keys.}

\item{new_colAnns}{Name of new column annotations in custom annotations file}

\item{feat_sets_order}{Logical, should we preserve feat_sets order?}
}
\value{
A list object specifying: ds, colAnns, feat_sets_name (name of custom analysis)
}
\description{
Get a subset of dataset list object from custom analysis table
}
