% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{get_file}
\alias{get_file}
\title{Search file names}
\usage{
get_file(file_list, regex)
}
\arguments{
\item{file_list}{A character vector of file names.}

\item{regex}{Regular expression pattern. See \code{\link[base]{regex}} for more info.}
}
\value{
The name(s) of file(s) that match the regex
}
\description{
Provide a list of file names and regular expression to get the file name
}
