% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_nth_part}
\alias{get_nth_part}
\title{Extracts the first (or nth) entry from each split.}
\usage{
get_nth_part(x, delimiter, entry_no)
}
\arguments{
\item{x}{A character vector.}

\item{delimiter}{A character to split by.}

\item{entry_no}{A number indicating which part of split to return.}
}
\value{
A character vector containing parts
}
\description{
Extracts the first (or nth) entry from each split.
}
\examples{
x <- c("2wk.1", "2wk.2", "2wk.3", "2wk.4", "2wk.5", "2wk.6", "2wk.7", "2wk.8", "2wk.9", "2wk.10")
# delimiter = "\\." # double slashes are escape characters for special characters like "."
# entry_no = 1L
get_nth_part(x, "\\\\.", "\\\\.")
# The return value is c("2wk", "2wk", "2wk", "2wk", "2wk", "2wk", "2wk", "2wk", "2wk", "2wk", "2wk", "2wk")
get_nth_part(x, "wk", 1)
}
