% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{get_out_dir}
\alias{get_out_dir}
\title{PDAC-specific - Customize output folder name depending on exclusion criteria and presence of row annotation 2}
\usage{
get_out_dir(
  current_dir,
  all_out_dirs = NULL,
  rowAnn2 = NA,
  EXC_HRD = F,
  EXC_NEO = F
)
}
\arguments{
\item{current_dir}{The name of the output directory}

\item{all_out_dirs}{A character vector of directory names, if the current_dir name exists, the algorithm will append "1", "2", and so on until the dir name is unique.}

\item{rowAnn2}{Name of row annotation 2 if applicable, otherwise NA.}

\item{EXC_HRD}{Logical, include "excl HRD" label?}

\item{EXC_NEO}{Logical, include "excl neo" label?}
}
\value{
New name of output folder
}
\description{
PDAC-specific - Customize output folder name depending on exclusion criteria and presence of row annotation 2
}
