% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_rowAnn_color_pal}
\alias{get_rowAnn_color_pal}
\title{Get color palette for dataset row annotations}
\usage{
get_rowAnn_color_pal(ds, rowAnns, var_colors = NULL)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn (required), colAnn)}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn.}

\item{var_colors}{A named vector with colors as values and annotations/groups as names.}
}
\value{
A list of 2 elements: 1) ds without NAs in rowAnn1, 2) rowAnns
}
\description{
Get color palette for dataset row annotations
}
