% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_at.least_n.vals}
\alias{has_at.least_n.vals}
\title{Checks whether a vector has at least a certain number of non NA/Nan/NULL values}
\usage{
has_at.least_n.vals(v, n = 0, row.or.col = 1)
}
\arguments{
\item{v}{A vector (any type)}

\item{n}{A threshold of non NA values}

\item{row.or.col}{A number indicating direction to apply function: either 1 for row-wise, 2 (default) for column-wise.}
}
\value{
A logical indicating whether the number of non-NA values is above the threshold
}
\description{
Checks whether a vector has at least a certain number of non NA/Nan/NULL values
}
