% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_less.than.eq.to_NA.thres}
\alias{has_less.than.eq.to_NA.thres}
\title{Checks which rows or columns have number of NAs less than or equal to NA threshold}
\usage{
has_less.than.eq.to_NA.thres(df, NA_thres, row.or.col = 1)
}
\arguments{
\item{df}{Data frame or matrix}

\item{NA_thres}{A number for NA threshold. 0 means that row/col has no NAs}

\item{row.or.col}{A number indicating direction to apply function: either 1 for row-wise, 2 (default) for column-wise.}
}
\value{
A logical vector indicating whether the number of non-NA values is above the threshold: number of NA values <= NA threshold
}
\description{
Checks which rows or columns have number of NAs less than or equal to NA threshold
}
