% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_w_mean.R
\name{impute_ds}
\alias{impute_ds}
\title{Impute numeric matrix in dataset object}
\usage{
impute_ds(ds, p = 5, add_to_name = "imputed")
}
\arguments{
\item{ds}{Dataset list object with vals, rowAnn, colAnn, name}

\item{p}{the percent around the mean to impute, default: p=5\%=5}

\item{add_to_name}{String to add to the ds$name for new imputed dataset, default: "imputed"}
}
\value{
Dataset object where NA in vals has been imputed with random values around mean (default is 5\%)
}
\description{
Impute numeric matrix in dataset object
}
