% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_w_mean.R
\name{impute_w_mean}
\alias{impute_w_mean}
\title{Functions defined in this file:
impute_w_mean
impute_w_mean_df
impute_ds
Imputes NA/missing values in a vector with random values plus/minus the percent around the mean of the values in vector}
\usage{
impute_w_mean(v, p = 5)
}
\arguments{
\item{v}{A numeric vector containing NAs}

\item{p}{the percent around the mean, e.g. p=5\%=5}
}
\value{
The input vector with NAs imputed.
}
\description{
Functions defined in this file:
impute_w_mean
impute_w_mean_df
impute_ds
Imputes NA/missing values in a vector with random values plus/minus the percent around the mean of the values in vector
}
