% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_w_mean.R
\name{impute_w_mean_df}
\alias{impute_w_mean_df}
\title{Imputes NA/missing values in a data frame with random values plus/minus the percent around the mean across rows/columns}
\usage{
impute_w_mean_df(df, row.or.col = 2, p = 5)
}
\arguments{
\item{df}{Numeric data frame or matrix}

\item{row.or.col}{Number indicating direction to apply function - either 1 (rows) or 2 (columns, default)}

\item{p}{the percent around the mean, e.g. p=5\%=5}
}
\value{
The input data frame with NAs imputed.
}
\description{
Imputes NA/missing values in a data frame with random values plus/minus the percent around the mean across rows/columns
}
