% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make.unique.2}
\alias{make.unique.2}
\title{R make.unique starting in 1}
\usage{
make.unique.2(x, sep = ".")
}
\arguments{
\item{x}{A numeric vector}

\item{sep}{Delimeter between value and number}
}
\description{
R make.unique starting in 1. Example: x,x.1,x.2,x.3 becomes x.1,x.2,x.3,x.4.
}
