% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC.pval.R
\name{make_FC.pval_df}
\alias{make_FC.pval_df}
\title{Find p-values and fold-change (FC) for each group}
\usage{
make_FC.pval_df(
  df,
  rowAnn_col = 1,
  rev = T,
  vs.other = T,
  group_name_sep = "/",
  FC.method = "divide",
  p.method = "wilcox"
)
}
\arguments{
\item{df}{A data frame - first column: rowAnnotation column with groups, the rest of the columns are values}

\item{rowAnn_col}{A column index (numeric) or name in df indicating which groups to stratify by}

\item{rev}{Logical indicating whether to reverse the order of the default FC (default is TRUE), ex. low-hig --> hig-low}

\item{vs.other}{logical, indicating whether to compare all groups to other, eg. high vs (low+intermed)}

\item{group_name_sep}{A character indicating how the labels for groups should be presented, "/" in low/high}

\item{FC.method}{Fold change computation method to use, either "divide" (for non-transformed values) or "subtract" (for log2-transformed values)}

\item{p.method}{Significance test to use, either "t.test" or "wilcox"}
}
\value{
A data frame with p-vals and fold-changes for each group (rowAnn) in each variable (column)
}
\description{
Find p-values and fold-change (FC) for each group specified in rowAnn_col across many columns of a data frame
}
