% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC.pval.R
\name{make_FC.pval_df_helper}
\alias{make_FC.pval_df_helper}
\title{Find p-values and fold-change (FC) for each group}
\usage{
make_FC.pval_df_helper(
  df3,
  rowAnn_col = 1,
  val_col = 2,
  rev = F,
  group_name_sep = "/",
  FC.method = "divide",
  p.method = "wilcox.test"
)
}
\arguments{
\item{df3}{A data frame with 2 columns: 1) row annotations (rowAnn), 2) numeric values}

\item{rowAnn_col}{Column in df3 with the groups}

\item{val_col}{Column in df3 with the numeric values}

\item{rev}{Logical indicating whether to reverse the order of the default FC (default is TRUE), ex. low-hig --> hig-low}

\item{group_name_sep}{A character indicating how the labels for groups should be presented, "/" in low/high}

\item{FC.method}{Fold change computation method to use, either "divide" (for non-transformed values) or "subtract" (for log2-transformed values)}

\item{p.method}{Significance test to use, either "t.test" or "wilcox.test"}
}
\value{
A data frame with p-values and fold-changes
}
\description{
Find p-values and fold-change (FC) for each group
}
\section{Output data frame}{

group    p.value Fold.change
1 intermed/high 0.03457640   0.2216307
2      low/high 0.02586801   0.5496807
3  low/intermed 0.65911963   2.4801647
}

