% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{means_of_groups}
\alias{means_of_groups}
\title{Calculates means of a vector with discrete levels}
\usage{
means_of_groups(vals, groups)
}
\arguments{
\item{vals}{A numeric vector of values}

\item{groups}{A character vector specifying the respective group of each value}
}
\value{
A data frame with 2 columns: 1) group name and 2) median value of group
}
\description{
Calculates means of a vector with discrete levels
}
