% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_t.test}
\alias{perform_t.test}
\title{Computes t-test between 2 groups}
\usage{
perform_t.test(vals, groups, ...)
}
\arguments{
\item{vals}{A numeric vector of values}

\item{groups}{A character vector specifying the respective group of each value}

\item{...}{Additional parameters passed to ?pairwise.t.test}
}
\value{
A number indicating p-value from statistical test
}
\description{
Computes t-test between 2 groups
}
