% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{perform_wilcox}
\alias{perform_wilcox}
\title{Computes Wilcoxon test between 2 groups}
\usage{
perform_wilcox(vals, groups, p.adjust.method = "none", paired = F, ...)
}
\arguments{
\item{vals}{A numeric vector of values}

\item{groups}{A character vector specifying the respective group of each value}

\item{...}{Additional parameters passed to ?pairwise.wilcox.test.}
}
\value{
A number indicating p-value from statistical test
}
\description{
Computes Wilcoxon test between 2 groups
}
