% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{plot_corrplotgg}
\alias{plot_corrplotgg}
\title{Creates correlation plot using ggplot2 package}
\usage{
plot_corrplotgg(
  mat,
  xlab = "",
  ylab = "",
  labels = "",
  corr_method = c("pairwise.complete.obs", "spearman"),
  pval_color = "white",
  grid.fill.color = "white",
  grid.line.color = "black",
  p_signif = "stars",
  circ_max = NULL,
  star_size = NULL,
  pal_brew = "RdBu",
  font_size = 15,
  line_size = 1,
  out_dir = ".",
  save.to.file = F
)
}
\arguments{
\item{mat}{Numeric dataframe or matrix, where columns will be correlated. Must not contain any NAs; use complete.cases(mat) to remove rows with NAs.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{corr_method}{A character vector of 2 that correspond to use and method parameters in \code{\link[stats]{cor}}.}

\item{pval_color}{The color of the significance stars or p-value text.}

\item{grid.fill.color}{The color for grid fill.}

\item{grid.line.color}{The color for grid line.}

\item{p_signif}{How p-values are represented. Either "stars" or "text".}

\item{circ_max}{The maximum size of circle within cells in grid.}

\item{star_size}{The size of star labels on correlation plots.}

\item{pal_brew}{RColorBrewer palette. See \code{\link[RColorBrewer]{display.brewer.all}} for all options. Note: Define PAL_BREWER global variable.}

\item{font_size}{The size of text labels on correlation plots. legend title. The size of legend text and plot title is font_size / 1.5. The size of legend text and plot subtitle is font_size / 3.}

\item{line_size}{The thickness of grid lines.}

\item{out_dir}{The output directory where the plot will be saved when save.to.file is TRUE, default is current working directory.}

\item{save.to.file}{If TRUE, save plot to file in out_dir. If FALSE, print to panel.}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Creates correlation plot using ggplot2 package
}
\seealso{
Other plotting: 
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
