% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_barplot.R
\name{plot_discrete_barplot}
\alias{plot_discrete_barplot}
\title{Plot discrete bar graph.}
\usage{
plot_discrete_barplot(
  df2,
  out_dir = ".",
  xlab = "",
  ylab = "",
  facet_by_var = F,
  legend_title = "",
  pal = NA,
  plot_title = "",
  pal_brew = "RdBu",
  pos = "stack",
  font_size = 20,
  line_size = 2,
  save.to.file = F
)
}
\arguments{
\item{df2}{A data frame with 3 columns: 1) ID = discrete values to group by (i.e. rowAnn1) = bars, 2) variables on x-axis = group of bars, 3) value.}

\item{out_dir}{The output directory where the plots will be saved, default is current working directory.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{facet_by_var}{If TRUE, facet groups of bars horizontally.}

\item{legend_title}{Title of legend.}

\item{pal}{Vector for colors of values.}

\item{plot_title}{Title of plot.}

\item{pal_brew}{RColorBrewer palette. See RColorBrewer::display.brewer.all() for all options. Note: Define PAL_BREWER global variable.}

\item{pos}{How bars should be stacked. Either "fill" (relative ratio, 100\% bar) or "stack". See position parameter in \code{\link[ggplot2]{geom_bar}}}

\item{font_size}{The size of axis title on plots. The size of plot subtitle and caption is font_size / 2. The size of legend text and x axis text is font_size / 3 and font_size / 1.5.}

\item{line_size}{The thickness of axis lines.}

\item{save.to.file}{If TRUE, save plot to file in out_dir. If FALSE, print to panel.}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Plot discrete bar graph.
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
