% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plots a heatmap}
\usage{
plot_heatmap(
  mat,
  ann_row = NA,
  ann_col = NA,
  ann_colors = NA,
  plot_title = "",
  out_dir = ".",
  labels = "",
  clust_row = F,
  clust_col = F,
  log10 = F,
  z_score = F,
  man_scale = T,
  man_scale_range = c(-2, 2),
  scale_func = F,
  col.or.row = 2,
  pheatmap_scale = "none",
  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean",
  clustering_method = "complete",
  show_colnames = T,
  show_rownames = F,
  pal_brew = "RdBu",
  border_color = NA,
  ...
)
}
\arguments{
\item{mat}{Numeric matrix or data frame directory.}

\item{ann_row}{A data frame that defines annotations or features for rows of the data matrix. The row names should the same as the row names of the data matrix.}

\item{ann_col}{A data frame that defines annotations or features for columns of the data matrix. The row names should the same as the column names of the data matrix.}

\item{ann_colors}{A list that specifies annotation colors (each element is a named vector, where names are unique groups in annotations), names of list are the annotation names.}

\item{out_dir}{The output directory where the plot will be saved, default is current working}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name.}

\item{clust_row}{Logical indicating whether to cluster rows}

\item{clust_col}{Logical indicating whether to cluster rows}

\item{log10}{Logical indicating whether to log10-transform values before plotting}

\item{z_score}{Logical indicating whether to plot z_scores}

\item{man_scale}{Logical indicating whether to manually scale each row/column using rescale() from scales package}

\item{man_scale_range}{Numeric vector of length 2 indicating desired range of values e.g. c(-4,4), applied when man_scale is TRUE.}

\item{scale_func}{Logical indicating whether to scale matrix using scale()}

\item{col.or.row}{numeric - 1 for row, 2 for column. Only applicable when z_score, scale_func, man_scale are TRUE.}

\item{pheatmap_scale}{String describing which direction around the mean to scale by. Accepts the same values ("none", "row", "column") as scale parameter in ?pheatmap}

\item{clustering_distance_rows}{Distance measure to be used for row clustering. Accepts the same values as method parameter in ?dist.}

\item{clustering_distance_cols}{Distance measure to be used for column clustering. Accepts the same values as method parameter in ?dist.}

\item{clustering_method}{The agglomeration/clustering method to be used. Accepts the same values as method parameter in ?hclust.}

\item{show_colnames}{Logical indicating whether to show column names on heatmap. See ?pheatmap}

\item{show_rownames}{Logical indicating whether to show row names on heatmap. See ?pheatmap}

\item{pal_brew}{RColorBrewer palette. See RColorBrewer::display.brewer.all() for all options. Note: Define PAL_BREWER global variable.}

\item{...}{Additional plotting parameters. Parameters passed to pheatmap, see ?pheatmap.}
}
\description{
Transforms a numeric data frame/matrix and plots corresponding heatmap using the pheatmap function and saves to file
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
