% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{plot_indiv_boxplot}
\alias{plot_indiv_boxplot}
\title{Creates single box plot.}
\usage{
plot_indiv_boxplot(
  df,
  labels = "Group",
  out_dir = ".",
  log10 = T,
  font_size = 25,
  line_size = 1.3,
  color_pal = NA,
  xlab = "",
  ylab = "value",
  rowAnns = c(NA, NA),
  alpha_dots = 0.8,
  alpha_box = 1,
  point_size = 2,
  jit_w = 0.1,
  pval.test = "wilcox.test",
  pval.label = "p.signif",
  trim_x = 3,
  save.to.file = T
)
}
\arguments{
\item{df}{2-3 columns. 1) Box or level, 2) Value 3) Dots (color)}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{log10}{Logical (TRUE/FALSE) indicating whether to log10-transform values before plotting.}

\item{font_size}{The size of axis.text. The size of axis.title and plot.title is font_size / 1.3 and font_size / 2, respectively. The size of plot.subtitle, legend.text, and legend.title is font_size/3.}

\item{line_size}{The thickness of axis lines.}

\item{color_pal}{A vector of colors to manually specify box colors.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{rowAnns}{Optional. A character vector of 1-2 column names in ds$rowAnn.}

\item{alpha_dots}{Transparency (alpha) of dots. Accepted values in range 0-1.}

\item{alpha_box}{Transparency (alpha) of boxes. Accepted values in range 0-1.}

\item{point_size}{Size of dots.}

\item{jit_w}{Jitter width of dots.}

\item{pval.test}{String corresponding to method parameter in \code{\link[ggpubr]{stat_compare_means}}. Allowed values are "t.test" and "wilcox.test".}

\item{pval.label}{p-values on box plots, either "p.signif" (stars), "p.format" (numeric), etc.}

\item{trim_x}{Number of characters in x-axis labels.}

\item{save.to.file}{If TRUE, save plot to file in out_dir. If FALSE, print to panel.}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Creates single box plot.
}
\examples{
# Reorder data frame so "box" column is first and value is second
plot_indiv_boxplot(ToothGrowth[,c("supp", "len")], save.to.file = F)
# color code dots is third column
plot_indiv_boxplot(ToothGrowth[,c("supp", "len", "dose")], save.to.file = F)
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
