% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrscatt.R
\name{plot_indiv_corrscatt}
\alias{plot_indiv_corrscatt}
\title{Plot all correlation scatter plots for pairwise variables}
\usage{
plot_indiv_corrscatt(
  df,
  rowAnn_col = 1,
  cor.method = "spearman",
  out_dir = ".",
  labels = ""
)
}
\arguments{
\item{df}{A data frame with the first column as discrete values to group by (i.e. rowAnn_col), and the rest of the columns are numeric variables to plot.}

\item{rowAnn_col}{A column index (numeric) or name in df indicating which groups to stratify by.}

\item{cor.method}{Method for correlation (one of "pearson","spearman","kendall").}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}
}
\description{
Plot all correlation scatter plots for pairwise variables
}
\section{Warning}{

This function will take a long time to run depending on the number of variables.
}

\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_het_barplot}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
