% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.R
\name{plot_indiv_paired}
\alias{plot_indiv_paired}
\title{Create paired plot}
\usage{
plot_indiv_paired(
  df,
  labels = "Group",
  out_dir = ".",
  font_size = 30,
  line_size = 1.3,
  color_pal = NA,
  xlab = "variable",
  ylab = "value",
  rowAnns = c(NA, NA),
  pval.test = "t.test",
  pval.label = "p.format",
  log10_y = F,
  save.to.file = T
)
}
\arguments{
\item{df}{A data frame with 3 columns: 1) id = cases or groups to pair by, 2) box = dif levels in rowAnn, 3) value}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{font_size}{The size of text labels on correlation plots. legend title. The size of legend text and plot title is font_size / 1.5. The size of legend text and plot subtitle is font_size / 3.}

\item{line_size}{The thickness of grid lines.}

\item{color_pal}{A named vector with colors, where names are box names (values in column 2 of df).}

\item{ylab}{Y axis label.}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn.}

\item{pval.test}{2-sample test to use (paired only for 2 groups). String corresponding to method parameter in \code{\link[ggpubr]{stat_compare_means}}. Allowed values are "t.test" and "wilcox.test".}

\item{pval.label}{p-value label. String corresponding to label parameter in \code{\link[ggpubr]{stat_compare_means}}. Allowed values are "p.signif" (stars) and "p.format" (number).}

\item{save.to.file}{If TRUE, save plot to file in out_dir. If FALSE, print to panel.}

\item{log10}{Logical indicating whether to log10-transform values before plotting}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Create paired plot
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_het_barplot}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
