% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{plot_overview_boxplot}
\alias{plot_overview_boxplot}
\title{Creates overview box plots.}
\usage{
plot_overview_boxplot(
  df3,
  out_dir = ".",
  labels = "",
  log10_y = F,
  lvl.colors = NA,
  pal_brew = "RdBu",
  legend.title = "Group",
  xlab = "variable",
  dot_color = "black",
  ylab = "value",
  font_size = 15,
  line_size = 1.3,
  save.to.file = T
)
}
\arguments{
\item{df3}{A data frame in long format with 3 columns: order must be: 1) variable to group by (groups of boxes) = factor/character, 2) levels within each group (boxes) = factor/character, 3) values = numeric}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{log10_y}{Logical (TRUE/FALSE) indicating whether to log10-transform y-axis.}

\item{lvl.colors}{A vector of colors to manually specify levels colors = box colors.}

\item{pal_brew}{RColorBrewer palette for variable colors if lvl.colors is NA. See RColorBrewer::display.brewer.all() for all options. Note: Define PAL_BREWER global variable.}

\item{legend.title}{Title for legend.}

\item{xlab}{X axis label.}

\item{dot_color}{A string indicating what color to make points, if NA, points will not be shown.}

\item{ylab}{Y axis label.}

\item{font_size}{The size of axis.title, axis.text, plot.title, legend.text and legend.title. The size of plot.subtitle is font_size / 2.}

\item{line_size}{The thickness of axis lines.}

\item{save.to.file}{If TRUE, save plot to file in out_dir. If FALSE, print to panel.}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Creates overview box plots.
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
