% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrscatt.R
\name{plot_overview_corr_scatt}
\alias{plot_overview_corr_scatt}
\title{Plot overview correlation scatter plot}
\usage{
plot_overview_corr_scatt(
  mat,
  out_dir = ".",
  labels = "",
  corr_method = "pearson",
  hist.col = "#92C5DE",
  save.to.file = T,
  compute_w_NA = F,
  ...
)
}
\arguments{
\item{mat}{Numeric data frame or matrix}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{corr_method}{Method for correlation (one of "pearson","spearman","kendall").}

\item{hist.col}{Histogram color on the diagonal.}

\item{save.to.file}{Logical indicating whether to save to out_dir (TRUE) or print to panel (FALSE).}

\item{compute_w_NA}{If TRUE, consider columns and rows with at least 7 values (default = FALSE).}

\item{...}{Additional parameters passed to \code{\link[psych]{pairs.panels}}. e.g. rug = T, scale = T (scale font by size of correlation), jiggle = T (jitter points).}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Plot overview correlation scatter plot
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_het_barplot}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
