% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_barplot.R
\name{plot_profile_barplot}
\alias{plot_profile_barplot}
\title{Plot profile bar graph}
\usage{
plot_profile_barplot(
  df2,
  pos = "stack",
  var_colors = NA,
  out_dir = ".",
  legend_title = "Group",
  labels = "",
  gradient_palette = "RdBu",
  line_size = 1,
  font_size = 10,
  save.to.file = F
)
}
\arguments{
\item{df2}{A long format data frame with 4 columns: 1) group (group of bars on x), 2) ID (individual bars on x), 3) variable (components within each bar), 4) value.}

\item{pos}{How bars should be stacked. Either "fill" (relative ratio, 100\% bar) or "stack". See position parameter in \code{\link[ggplot2]{geom_bar}}}

\item{var_colors}{A vector of colors to manually specify variable colors.}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{gradient_palette}{RColorBrewer palette for variable colors if var_colors is NA. See RColorBrewer::display.brewer.all() for all options.}

\item{line_size}{The thickness of axis lines.}

\item{font_size}{The size of axis title on plots. The size of plot subtitle and caption is font_size / 2. The size of legend text and x axis text is font_size / 3 and font_size / 1.5.}

\item{save.to.file}{If TRUE, save plot to file in out_dir. If FALSE, print to panel.}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Plot profile bar graph
}
\section{Example of input data frame}{

group         ID        variable     value
1       low 1_21043      CD3  273.2400
2       low 1_36312      CD3  700.2100
3       low 1_37265      CD3  931.1133
4       low 1_39924      CD3 1503.2325
}

\seealso{
Other plotting: 
\code{\link{plot_corrplot}()},
\code{\link{plot_corrplotgg}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_het_barplot}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
