% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{plot_surv_curve}
\alias{plot_surv_curve}
\title{Plot survival curve (Kaplan Meier)}
\usage{
plot_surv_curve(df1, label = "", out_dir = ".", save.to.file = T)
}
\arguments{
\item{df1}{A data frame with 3 columns: time, status (censoring), col (variable to stratify by)}

\item{label}{A string label for file names and plot title; could be strata name}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{save.to.file}{A logical indicating whether to save to out_dir (T) or print to panel (F)}
}
\description{
Plot survival curve (Kaplan Meier)
}
