% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_unique_elements_count}
\alias{print_unique_elements_count}
\title{Prints the number of unique elements in each column}
\usage{
print_unique_elements_count(msg, df, column_names)
}
\arguments{
\item{msg}{String description of data frame}

\item{df}{A data frame}

\item{column_names}{A character vector of which columns to search}
}
\description{
Prints the number of unique elements in each column
}
