% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pval_to_stars}
\alias{pval_to_stars}
\title{Convert p-values to stars}
\usage{
pval_to_stars(pvals)
}
\arguments{
\item{pvals}{A numeric vector of pvals}
}
\value{
A character vector with stars representing p-values
}
\description{
Convert p-values to stars according to map legend p <= 0.001'\emph{\emph{\strong{', 0.001 '}}', 0.01 '**', 0.05 '}'
}
