% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{read_file}
\alias{read_file}
\title{Read tabular data}
\usage{
read_file(filename, sep = "\\t")
}
\arguments{
\item{filename}{String with path to file}

\item{sep}{The field separator character. Values on each line of the file are separated by this character.}

\item{...}{Additional parameters passed to \code{\link[utils]{read.table}}}
}
\value{
The data frame stored in file
}
\description{
Read tabular data - file input can only be tab-delimited text or csv.
}
