% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reform_ann_df}
\alias{reform_ann_df}
\title{Subset data frame by columns}
\usage{
reform_ann_df(ann_df, anns)
}
\arguments{
\item{ann_df}{A data frame with row names}

\item{anns}{Column names in ann_df desired in output}
}
\value{
The data frame, ann_df, with columns subsetted to names specified by anns or NA if
}
\description{
Prevents subset of data frame with one column from becoming a vector.
}
\examples{
reform_ann_df(DNase, "conc")
}
