% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers outside of 1st and 3rd quartile by making the values NA}
\usage{
remove_outliers(x)
}
\arguments{
\item{x}{A numeric vector, can contain NA.}
}
\value{
The numeric vector without outliers (points outside 1st/3rd quartiles)
}
\description{
Remove outliers outside of 1st and 3rd quartile by making the values NA
}
