% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers2}
\alias{remove_outliers2}
\title{Removes outliers using boxplot statistics}
\usage{
remove_outliers2(x)
}
\arguments{
\item{x}{A numeric vector, can contain NA.}
}
\value{
The numeric vector without outliers
}
\description{
Removes outliers using boxplot statistics
}
