% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers_df}
\alias{remove_outliers_df}
\title{Remove outliers outside of 1st and 3rd quartile by making the values NA}
\usage{
remove_outliers_df(df, row.or.col = 2)
}
\arguments{
\item{df}{Numeric data frame or matrix}

\item{row.or.col}{A number indicating direction to apply function: either 1 for row-wise, 2 (default) for column-wise.}
}
\value{
The data frame without outliers
}
\description{
Remove outliers outside of 1st and 3rd quartile by making the values NA
}
