% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_column}
\alias{rename_column}
\title{Rename a column in a data frame.}
\usage{
rename_column(df, current_col, new_name)
}
\arguments{
\item{df}{A data frame.}

\item{current_col}{Column name in df that you wish to rename.}

\item{new_name}{New column name.}
}
\value{
The dataframe with the column renamed.
}
\description{
Rename a column in a data frame.
}
