% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comparisons.R
\name{run_comparison}
\alias{run_comparison}
\title{Run analysis in streamlined approach}
\usage{
run_comparison(
  ds,
  rowAnns,
  colAnns = NA,
  output_folder = ".",
  ds.imp = NULL,
  feat_sets = NULL,
  var_colors = NULL,
  gradient_palette = "RdBu",
  corr_method = "pearson",
  pval.test = "t.test",
  pval.label = "p.signif",
  FC.method = "divide",
  paired_analysis_column = NA,
  make.QC.param = F,
  make.QC.feature = F,
  discrete_stacked_params = NULL,
  make.het.plot = F,
  make.indiv.boxplot = F,
  make.overview.boxplot = F,
  make.heatmap = F,
  make.corrplot = F,
  make.overview.corrscatt = F,
  make.indiv.corrscatt = F,
  make.barplot = F,
  make.FC.pval.plot = F
)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn. c(MainComparison, Subgroup)}

\item{colAnns}{A character vector of 1-2 column names in ds$colAnn. c(Parameter, Feature/Stain/Gene)}

\item{output_folder}{The main output folder for all custom analysis plots and boxplots for by.parameter and by.feature analysis}

\item{ds.imp}{A dataset object similar to ds with imputed or another transformation values.}

\item{feat_sets}{A list of 2 data frames for feature sets and parameters.}

\item{var_colors}{A named vector with colors as values and annotations/groups as names.}

\item{gradient_palette}{RColorBrewer palette name for gradients (e.g. heatmap, correlation plots). See RColorBrewer::display.brewer.all() for all options.}

\item{corr_method}{Method for correlation (one of "pearson","spearman","kendall").}

\item{pval.test}{Which two-samples testing should be used? String corresponding to "method" parameter in \code{\link[ggpubr]{stat_compare_means}}. Allowed values are "t.test" and "wilcox.test".}

\item{pval.label}{How to display p-values? String corresponding to "label" parameter in \code{\link[ggpubr]{stat_compare_means}}. Allowed values are "p.signif" (stars) and "p.format" (number).}

\item{FC.method}{Fold change computation method to use, either "divide" (for non-transformed values) or "subtract" (for log2-transformed values)}

\item{paired_analysis_column}{column name in ds$rowAnn to create paired analysis plots for, e.g. PatientID if ds is data for all cores}

\item{make.QC.param, make.QC.feature, make.het.plot, make.indiv.boxplot, make.overview.boxplot, make.heatmap, make.corrplot, make.overview.corrscatt, make.indiv.corrscatt, make.barplot, make.FC.pval.plot}{Logicals (TRUE/FALSE) indicating whether to make these plots. Note: make.indiv.corrscatt = T takes a long time.}

\item{discrete_stacked_params}{parameter names to search for in colAnn1 to make discrete stacked barplots, e.g. "Het.Score"}

\item{run}{A one row data frame or list object with logicals for what to run, names: make.boxplots, make.paired.boxplots, make.heatmaps, make.surv.curve}
}
\description{
Create the plots for each comparison. Run the same analysis on dataset object (ds parameter) and transformed values (ds.imp parameter, optional).
}
