% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_comparisons.R
\name{run_comparisons}
\alias{run_comparisons}
\title{Run boxplot and heatmap analysis in streamlined approach}
\usage{
run_comparisons(
  ds,
  rowAnns,
  colAnns = NA,
  heatmap_folder = ".",
  boxplot_folder = ".",
  run = list(),
  ds.imp = NULL,
  customAn = NULL
)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn.}

\item{colAnns}{A character vector of 1-2 column names in ds$colAnn.}

\item{heatmap_folder}{The main output folder for heatmaps and all other plots (for custom analysis and within parameter analysis)}

\item{boxplot_folder}{The main output folder for all boxplots (individual, overview, paired).}

\item{run}{A one row data frame or list object with logicals for what to run, names: make.boxplots, make.paired.boxplots, make.heatmaps, make.surv.curve}

\item{ds.imp}{A dataset object similar to ds with imputed or another transformation values.}

\item{customAn}{A list of 2 data frames for plotting specific rows and columns. @seealso customAn read in using \code{\link{import_custom_analysis_file}}, see description of file.}
}
\description{
Create the plots for each comparison. Run the same analysis on dataset object (ds parameter) and transformed values (ds.imp parameter, optional).
}
