% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{run_corrplot_analysis}
\alias{run_corrplot_analysis}
\title{Run correlation plot analysis}
\usage{
run_corrplot_analysis(
  df,
  rowAnn_col = 1,
  out_dir = ".",
  labels = "",
  gradient_palette = "RdBu",
  plot_gg = T,
  corr_method = c("pairwise.complete.obs", "spearman"),
  pval.label = "p.signif"
)
}
\arguments{
\item{df}{A data frame - first column: rowAnnotation column with groups, the rest of the columns are values.}

\item{rowAnn_col}{A column index (numeric) or name in df indicating which groups to stratify by.}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{gradient_palette}{RColorBrewer palette. See RColorBrewer::display.brewer.all() for all options.}

\item{plot_gg}{Should we use the function \code{link{plot_corrplotgg}} from this package? Otherwise, it uses \code{link{plot_corrplot}}.}

\item{corr_method}{A character vector of 2 that correspond to "use" and "method" parameters in \code{\link[stats]{cor}}. The first defaults to "pairwise.complete.obs" and second should be one of "pearson","spearman","kendall".}

\item{pval.label}{How to display p-values? Allowed values are "p.signif" (stars) and "p.format" (number).}
}
\description{
Makes multiple correlation plots in 1 PDF file, each group and all values (total plots >= 2)
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_corrplot}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_het_barplot}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()}
}
\concept{plotting}
