% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_barplot.R
\name{run_discrete_barplot_analysis}
\alias{run_discrete_barplot_analysis}
\title{Makes discrete box plots for each parameter in column annotation}
\usage{
run_discrete_barplot_analysis(
  ds,
  rowAnn1,
  colAnns,
  parameters = "",
  out_dir = "."
)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnn1}{A column index (numeric) or name in ds$rowAnn indicating which groups to stratify by.}

\item{colAnns}{A character vector of 1-2 column names in ds$colAnn.}

\item{parameters}{Group names or parameters in colAnns\link{1} that required its own barplot}

\item{out_dir}{The output directory where the plots will be saved, default is current working directory.}
}
\description{
Makes discrete box plots for each parameter in column annotation
}
