% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hourglass.R
\name{run_from_excel}
\alias{run_from_excel}
\title{Run Hourglass from single Excel file input.}
\usage{
run_from_excel(xl_file)
}
\arguments{
\item{xl_file}{The path to an Excel file with 4 worksheets: Colors, Comparisons, FeatureSets, FeatureParameters. See documentation for more information.}
}
\description{
Run Hourglass according to user specifications indicated in Excel file input.
}
