% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{run_heatmap_analysis}
\alias{run_heatmap_analysis}
\title{Runs main analysis for all parameter combos specified in colAnns and creates plots.}
\usage{
run_heatmap_analysis(
  ds,
  rowAnns = 1,
  colAnns = NA,
  heatmap_folder = ".",
  customAn = NULL
)
}
\description{
Runs main analysis for all parameter combos specified in colAnns and creates plots.
}
