% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het_barplot.R
\name{run_het_analysis}
\alias{run_het_analysis}
\title{Runs heterogeneity analysis for custom comparison, look at expression levels within samples of same patient}
\usage{
run_het_analysis(ds, rowAnn1, pID = 1, out_dir = ".", var_colors = NULL)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnn1}{Column in ds$vals and was created in ds$rowAnn as a CustomComparison}

\item{pID}{short for patient ID, column name in ds$rowAnn to count samples for}

\item{out_dir}{The output folder path}

\item{var_colors}{A named vector with colors as values and annotations/groups as names.}
}
\description{
Runs heterogeneity analysis for custom comparison, look at expression levels within samples of same patient
}
