% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired.R
\name{run_paired_analysis}
\alias{run_paired_analysis}
\title{Makes multiple correlation plots in 1 PDF file}
\usage{
run_paired_analysis(ds, rowAnns, colAnns = NA, out_dir = ".")
}
\arguments{
\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}
}
\description{
Makes multiple correlation plots in 1 PDF file
}
