% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{save_table}
\alias{save_table}
\title{Print comparison data (MainComparison) + sample/patient ID as table to csv file.}
\usage{
save_table(ds, rowAnns, out_dir = ".")
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn. c(MainComparison, Subgroup)}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}
}
\description{
Print comparison data (MainComparison) + sample/patient ID as table to csv file.
}
