% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sep_param_by_levels}
\alias{sep_param_by_levels}
\title{Make a custom row annotation column}
\usage{
sep_param_by_levels(ds, custom_name)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{custom_name}{Either a column name in ds$rowAnn or a column in ds$vals.}
}
\value{
A list of 2 elements: 1) rowAnn1 = new row annotation column name, 2) rowAnn = new ds$rowAnn with the rowAnn1 column
}
\description{
Make new rowAnn column of expression level (low, intermediate, high) of a specific stain/parameter (i.e. any column in the data matrix ds$vals) or ds$rowAnn column
}
