% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_dataframe}
\alias{sort_dataframe}
\title{Sort a dataframe by the contents of its row (sort columnwise) or columns (sort row-wise)}
\usage{
sort_dataframe(df, row.or.col = "row", by = 1)
}
\arguments{
\item{df}{A dataframe or matrix}

\item{row.or.col}{Either "row" or "column" specify what is being sorted}

\item{by}{The name of the column/row to sort by.}
}
\value{
A data frame sorted to specified rows/columns order
}
\description{
Sort a dataframe by the contents of its row (sort columnwise) or columns (sort row-wise)
}
