% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort_dataset}
\alias{sort_dataset}
\title{Sort dataset list object to specified rows and/or columns order}
\usage{
sort_dataset(ds, row_order = NULL, col_order = NULL)
}
\arguments{
\item{ds}{A dataset object (a list with vals (required), rowAnn, colAnn)}

\item{row_order}{Numeric, logical or character vectors of row order. Note: character for named rows only}

\item{col_order}{Numeric, logical or character vectors of column order. Note: character for named columns only}
}
\value{
A dataset sorted to specified rows/columns order
}
\description{
Sort dataset list object to specified rows and/or columns order
}
