% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_dataframe}
\alias{subset_dataframe}
\title{Functions defined in this file:
subset_dataframe
subset_dataset
subset_by_filters
evaluate_filter
Subset data frame to the rows and/or columns specified}
\usage{
subset_dataframe(df, rows_to_keep = NULL, cols_to_keep = NULL)
}
\arguments{
\item{df}{The data frame or matrix to subset}

\item{rows_to_keep}{Numeric, logical or character vectors of rows to keep in subset. Note: character for named rows only}

\item{cols_to_keep}{Numeric, logical or character vectors of columns to keep in subset. Note: character for named columns only}
}
\value{
A data frame subsetted to specified rows/columns.
}
\description{
Functions defined in this file:
subset_dataframe
subset_dataset
subset_by_filters
evaluate_filter
Subset data frame to the rows and/or columns specified
}
