% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_dataset}
\alias{subset_dataset}
\title{Subset dataset list object to the rows and/or columns specified}
\usage{
subset_dataset(ds, rows_to_keep = NULL, cols_to_keep = NULL)
}
\arguments{
\item{ds}{A dataset object (a list with any of vals, rowAnn, colAnn)}

\item{rows_to_keep}{Numeric, logical or character vectors of rows to keep in subset. Note: character for named rows only}

\item{cols_to_keep}{Numeric, logical or character vectors of columns to keep in subset. Note: character for named columns only}
}
\value{
A dataset subsetted to specified rows/columns
}
\description{
Subset dataset list object to the rows and/or columns specified
}
