% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_sets.R
\name{subset_feat_sets_ds}
\alias{subset_feat_sets_ds}
\title{Subset dataset for custom analysis}
\usage{
subset_feat_sets_ds(
  ds,
  feat_sets,
  i,
  colAnns,
  std.or.alt = "Standard",
  feat_sets_order = T
)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{feat_sets}{A list of 2 data frames for plotting specific rows and columns.}

\item{i}{Row number in feat_sets$sets.}

\item{colAnns}{A character vector of 1-2 column names in ds$colAnn. c(Parameter, Feature/Stain/Gene)}

\item{std.or.alt}{string, specifies to use either "Standard" or "Alternative" parameters.}

\item{feat_sets_order}{Logical, should we preserve feat_sets order?}
}
\value{
A list object specifying: ds, colAnns, feat_sets_name (name of custom analysis)
}
\description{
Get a subset of dataset list object from custom analysis table
}
