% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_PDAC_rowAnn}
\alias{subset_PDAC_rowAnn}
\title{Subset PDAC-specific row annotation table based on selection criteria, provided as arguments.}
\usage{
subset_PDAC_rowAnn(
  rowAnn,
  EXC_HRD = F,
  EXC_NEO = F,
  LIMIT_TO_PANC = F,
  LIMIT_TO_PDAC = F
)
}
\arguments{
\item{rowAnn}{where row names are unique ids}

\item{EXC_HRD}{A logical, T will remove all HRD cases}

\item{LIMIT_TO_PANC}{A logical, T removes all non-pancreas organ controls (and missing cores) in Pancreas.tissue}

\item{LIMIT_TO_PDAC}{A logical, T positively selects for PDAC tissue in Pancreas.tissue}

\item{EXC_neo}{A logical, T will remove all neo-adjuvant cases}
}
\value{
A character vector of rownames of rowAnn to keep in analysis
}
\description{
Subset PDAC-specific row annotation table based on selection criteria, provided as arguments.
}
