% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trim_each_part}
\alias{trim_each_part}
\title{Abbreviate elements}
\usage{
trim_each_part(v, split = ",", trim_x = 3, paste_back = T)
}
\arguments{
\item{v}{A character vector.}

\item{split}{Character delimiter to split by. Note: escape characters still apply, e.g. for ".", split = "\\."}

\item{trim_x}{Number, indicating the number of characters for each part (that is, length of truncated output string).}

\item{paste_back}{Logical - paste the split elements back together?}
}
\value{
The abbreviated charactere elements
}
\description{
Split string elements in character vector or words in a string and abbreviate to specified number of characters.
}
\examples{
trim_each_part("abbreviate this", sep = " ")
z <- paste(rep("high", 4), rep(c("low", "intermediate"), each = 2), sep = "/")
# [1] "high/low"          "high/low"          "high/intermediate" "high/intermediate"
trim_each_part(paste(x, y, sep = "/"), split = "/")
# [1] "hig/low" "hig/low" "hig/int" "hig/int"
trim_each_part(z, split = "/", trim_x = 2)
# [1] "hi/lo" "hi/lo" "hi/in" "hi/in"
trim_each_part(z, split = "/", paste_back = F)
# [1] "hig" "low" "hig" "low" "hig" "int" "hig" "int"
}
