% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colors.R
\name{get_rowAnn_color_pal}
\alias{get_rowAnn_color_pal}
\title{Get color palette for dataset row annotations}
\usage{
get_rowAnn_color_pal(ds, rowAnns)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn (required), colAnn)}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn.}
}
\value{
A list of 2 elements: 1) ds without NAs in rowAnn\link{1}, 2) rowAnns
}
\description{
Get color palette for dataset row annotations
}
\note{
Specify PAL_SCAFF in global environment, named vector of current groups (preset).
}
