% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_files.R
\name{import_customAn_file}
\alias{import_customAn_file}
\title{Read in custom analysis file}
\usage{
import_customAn_file(file_xl)
}
\arguments{
\item{file_xl}{The name of file path to an Excel file for custom analyses}
}
\value{
List of 2 data frames: 1) keys, 2) values
}
\description{
Read in custom analysis file
}
