% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC.pval.R
\name{make_FC.pval_plot}
\alias{make_FC.pval_plot}
\title{Plot FC-p value heatmap}
\usage{
make_FC.pval_plot(
  df,
  x_lab = "",
  y_lab = "",
  plot_title = "",
  out_dir = ".",
  p_signif = "stars",
  pal_brew = "RdBu",
  group_name_sep = "/",
  trim_x = 3,
  pval_size = 8,
  pval_color = "white",
  log2FC = F,
  apply_scale_colFC = T,
  x_axis_angle = 0,
  save.to.file = F,
  font_size = 10,
  line_size = 1
)
}
\arguments{
\item{df}{A data frame with these exact columns 1) group (ie. group comparisons on x-axis), 2) p.value, 3) Fold.change, 4) Var (y-axis), created using \code{\link{make_p.val.FC_df}}.}

\item{plot_title}{Title of plot.}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}

\item{pal_brew}{RColorBrewer palette for fold-change values. See RColorBrewer::display.brewer.all() for all options. Note: Define PAL_BREWER global variable.}

\item{group_name_sep}{A character indicating how the labels for groups should be presented, "/" in low/high}

\item{trim_x}{Number, indicating the number of characters for each part (that is, length of truncated output string).}

\item{pval_size}{Size of p-values.}

\item{pval_color}{Color of p-values.}

\item{log2FC}{Logical (TRUE/FALSE). Should log2 transformation be applied to Fold.change column before plotting?}

\item{apply_scale_colFC}{Logical (TRUE/FALSE). Should scale be applied to FC for each group, ie. column-wise?}

\item{x_axis_angle}{Angle of the x-axis label. Default is 0 (horizontal), 90 is vertical.}

\item{save.to.file}{If TRUE, save plot to file in out_dir. If FALSE, print to panel.}

\item{font_size}{The size of text labels plot. legend title. The size of plot title, axis text, legend text is font_size. The size of plot subtitle is font_size / 1.5.}

\item{line_size}{The thickness of grid lines.}

\item{xlab}{X axis label.}

\item{ylab}{Y axis label.}

\item{p.signif}{How p-values should be plotted. Either "stars" or "number".}
}
\value{
Plot object if save.to.file is FALSE.
}
\description{
Plot grid of FC values as fill (with color scale) and p-value as stars/numbers in the centre of each tile.
}
\examples{

}
