% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplot.R
\name{plot_corrplot}
\alias{plot_corrplot}
\title{Creates correlation plot using corrplot package}
\usage{
plot_corrplot(
  mat,
  labels = "",
  text_size = 0.5,
  pch_size = 0.5,
  pal_brew = "RdBu",
  corr_method = c("pairwise.complete.obs", "spearman")
)
}
\arguments{
\item{mat}{Numeric data frame or matrix, where columns will be correlated. Must not contain any NAs; use complete.cases(mat) to remove rows with NAs.}

\item{labels}{A character vector of at least length 1 that will be collapsed for file name/plot titles.}

\item{text_size}{The size of text labels on correlation plots. Color legend size is text_size/1.5.}

\item{pch_size}{The size of star labels.}

\item{pal_brew}{RColorBrewer palette. See \code{\link[RColorBrewer]{display.brewer.all}} for all options. Note: Define PAL_BREWER global variable.}

\item{corr_method}{Method for correlation (one of "pearson","spearman","kendall")}
}
\value{
Plot object.
}
\description{
Creates correlation plot using corrplot package
}
\seealso{
Other plotting: 
\code{\link{plot_corrplotgg}()},
\code{\link{plot_discrete_barplot}()},
\code{\link{plot_heatmap}()},
\code{\link{plot_indiv_boxplot}()},
\code{\link{plot_indiv_corrscatt}()},
\code{\link{plot_indiv_paired}()},
\code{\link{plot_overview_boxplot}()},
\code{\link{plot_overview_corr_scatt}()},
\code{\link{plot_profile_barplot}()},
\code{\link{run_corrplot_analysis}()}
}
\concept{plotting}
