% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{run_boxplot_analysis}
\alias{run_boxplot_analysis}
\title{Creates multiple boxplots in a folder in streamlined approach.}
\usage{
run_boxplot_analysis(
  ds,
  rowAnns,
  colAnns = NA,
  out_dir = ".",
  make.overview.boxplot = T,
  make.indiv.boxplot = T
)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnns}{A character vector of 1-2 column names in ds$rowAnn. c(MainComparison, Subgroup)}

\item{colAnns}{A character vector of 1-2 column names in ds$colAnn. c(Parameter, Feature/Stain/Gene)}

\item{out_dir}{The output directory where the plots will be saved, default is current working directory.}

\item{make.overview.boxplot, make.indiv.boxplot}{Logical indicating which types of boxplots to create}
}
\description{
Create both individual boxplots (in 1 PDF file), and overview boxplots in seperate folders for each unique colAnn1 group. The boxes are specified by rowAnn1 and dots are colored by rowAnn2 (optional).
}
