% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_barplot.R
\name{run_discrete_barplot_analysis}
\alias{run_discrete_barplot_analysis}
\title{Makes discrete box plots for each parameter in column annotation}
\usage{
run_discrete_barplot_analysis(
  ds,
  rowAnn1 = 2,
  colAnns = NA,
  parameters = "",
  out_dir = ".",
  gradient_palette = "RdBu"
)
}
\arguments{
\item{ds}{A dataset object (a list with vals, rowAnn, colAnn, comparison, name).}

\item{rowAnn1}{A column index (numeric) or name in ds$rowAnn indicating which groups to stratify by.}

\item{colAnns}{A character vector of 1-2 column names in ds$colAnn. c(Parameter, Feature/Stain/Gene)}

\item{parameters}{Group names or parameters in colAnn1 that required its own barplot}

\item{gradient_palette}{RColorBrewer palette name for gradients (e.g. heatmap, correlation plots). See RColorBrewer::display.brewer.all() for all options.}
}
\description{
Makes discrete box plots for each parameter in column annotation
}
