% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hourglass.R
\name{run_Hourglass}
\alias{run_Hourglass}
\title{Run Hourglass.}
\usage{
run_Hourglass(
  comparisons,
  var_colors,
  feat_sets,
  main_folder = ".",
  datasets = NULL,
  keep_column_colAnn = "Keep.In.Analysis"
)
}
\arguments{
\item{comparisons}{A data frame containing comparisons to run compatible with run_hourglass function (i.e. column names are options, rows are each comparison to run). See ?get_comparisons for more info.}

\item{var_colors}{List of colors, where elements are hex codes and element names are rowAnn variables. e.g. list("Tumour"="#2f4f4Fff", "Stroma"="#d2691eff") See ?get_colors for more info.}

\item{feat_sets}{A list of 2 data frames for plotting specific rows and columns. See ?get_feat_sets for more info.}

\item{main_folder}{The output directory for Hourglass analysis, default is working directory.}

\item{datasets}{Optional. List of 2 elements 1) sample, 2) patient, which are the dataset objects for BySample and ByPatient analysis respectively.}

\item{keep_column_colAnn}{Optional. Column name in colAnn of which columns to keep in vals, important for QC plots}
}
\description{
Run Hourglass according to user specifications. See documentation for format of input parameters.
}
