% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{run_surv_analysis}
\alias{run_surv_analysis}
\title{Plot survival curves iteratively (Kaplan Meier).}
\usage{
run_surv_analysis(
  comparisons,
  patients,
  surv_time_column = "",
  surv_status_column = "",
  out_dir = ".",
  surv_folder = "Survival"
)
}
\arguments{
\item{comparisons}{A data frame containing comparisons to run compatible with run_hourglass function (i.e. column names are options, rows are each comparison to run). See ?get_comparisons for more info.}

\item{patients}{The dataset object for patients.}

\item{surv_time_column}{The column name in patients$rowAnn for overall survival time or disease free survival time.}

\item{surv_status_column}{The column name in patients$rowAnn indicating event or censoring status (dead=0, alive=1).}

\item{out_dir}{The output directory where the plots will be saved in a new folder called "Survival", default is current working directory.}

\item{surv_folder}{The subdirectory in out_dir where the survival plots will be saved, default is "Survival".}
}
\description{
Plot survival curves iteratively (Kaplan Meier).
}
