% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{run_surv_analysis}
\alias{run_surv_analysis}
\title{Functions defined in this file:
run_surv_analysis
plot_surv_curve
Plot survival curves (Kaplan Meier)}
\usage{
run_surv_analysis(df2, label = "", out_dir = ".")
}
\arguments{
\item{df2}{A data frame with 3 columns: time, status (censoring), col (variable to stratify by)}

\item{label}{A string label for file names and plot title; could be strata name}

\item{out_dir}{The output directory where the plot will be saved, default is current working directory.}
}
\description{
Functions defined in this file:
run_surv_analysis
plot_surv_curve
Plot survival curves (Kaplan Meier)
}
