% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_hourglass.R
\name{test_hourglass}
\alias{test_hourglass}
\title{Test hourglass R package from py interface.}
\usage{
test_hourglass(out_dir = ".", filename = "test_iris")
}
\arguments{
\item{out_dir}{The output directory for a file.}

\item{filename}{The name of the output csv file (minus the .csv extension).}
}
\description{
Used to test hourglass from py interface by saving the in-built iris dataset to a csv file in specified output directory.
}
